app/requirements.txt
Flask==0.12.2
Flask-Limiter==0.9.5.1
redis==2.10.6
marshmallow==2.13.6
python-logstash-async==1.2.0
bcrypt==3.1.4
twilio==6.8.1
pyjwt==1.5.3
gevent==1.2.2
gunicorn==19.7.1